+++ frontend/app/components/Homescreen/HomeCard/HomeCard.jsx
